Move all related logic into the store.
